+++ frontend/src/App.css
@@ -3,10 +3,11 @@ body {
   padding: 0;
   display: flex;
   justify-content: center;
-  align-items: center;
+  /* CRAZY MARGIN TODO */
+  margin-top: 5em;
   min-height: 100vh;
   height: auto;
-  background-color: #D3D3D3;
+  background-color: #f3f4f4;
   color: black;
   font-family: Arial, sans-serif;
   overflow-y: auto;
@@ -15,4 +16,10 @@ body {
 h1 {
   text-align: center;
   margin-bottom: 20px;
+}
+
+* {
+  padding: 0;
+  margin: 0;
+  box-sizing: border-box;
 }
