+++ frontend/src/components/dashboard/TopicsList/TopicsList.module.css
@@ -1,6 +1,4 @@
 .megaContainer {
-  padding-top: 10%;
-  padding-bottom: 5%;
 }
 
 .problem-container {
