+++ frontend/src/components/headers/PrimaryHeader/LoggedInHeader/LoggedInHeader.module.css
@@ -1,94 +1,90 @@
 .header {
-    position: fixed;
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 20vh; 
-    z-index: 1000; 
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    background-color: #161616;
-    padding: 0 4rem; 
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100%;
+  z-index: 1000;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  background-color: #161616;
+  padding: 1em 4rem;
 }
 
 .logo {
-    max-height: 80%;
+  max-width: 60%;
 }
 
 .userInfo {
-    display: flex;
-    flex-direction: column;
-    align-items: flex-end;
-    padding-right: 8rem;
+  display: flex;
+  flex-direction: column;
+  align-items: flex-end;
+  padding-right: 8rem;
 }
 
 .username {
-    color: #ffffff;
-    margin-bottom: 0.5rem;
+  color: #ffffff;
+  margin-bottom: 0.5rem;
 }
 
 .signout {
-    padding: 1rem 2rem;
-    border: solid;
-    border-radius: 1rem;
-    border-color: #ffffff;
-    border-width: 1px;
-    background-color: #161616;
-    color: #ffffff;
-    cursor: pointer;
-    margin-right: 1.25rem;
+  padding: 1rem 2rem;
+  border: solid;
+  border-radius: 1rem;
+  border-color: #ffffff;
+  border-width: 1px;
+  background-color: #161616;
+  color: #ffffff;
+  cursor: pointer;
+  margin-right: 1.25rem;
 }
 
 /* Changes for smaller screens */
 @media (max-width: 768px) {
-    .header {
-        flex-direction: column;
-        align-items: center;
-        padding: 1rem 2rem; 
-        height: auto; 
-    }
+  .header {
+    flex-direction: column;
+    align-items: center;
+    padding: 1rem 2rem;
+    height: auto;
+  }
 
-    .logo {
-        max-height: 60%;
-        margin-bottom: 1rem;
-    }
+  .logo {
+    max-height: 60%;
+    margin-bottom: 1rem;
+  }
 
-    .userInfo {
-        flex-direction: row;
-        justify-content: space-between;
-        align-items: center;
-        width: 100%;
-        padding-right: 0;
-    }
+  .userInfo {
+    flex-direction: row;
+    justify-content: space-between;
+    align-items: center;
+    width: 100%;
+    padding-right: 0;
+  }
 
-    .username {
-        margin-right: 1rem; 
-    }
+  .username {
+    margin-right: 1rem;
+  }
 
-    .signout {
-        padding: 0.5rem 1rem;
-    }
+  .signout {
+    padding: 0.5rem 1rem;
+  }
 }
 
+@media (max-height: 580px) {
+  .userInfo {
+    font-size: 0.5em;
+  }
 
-@media (max-height: 580px){
+  .userInfo .username {
+    margin-right: 1rem;
+  }
 
-    .userInfo {
-        font-size: 0.5em;
-    }
+  .userInfo .signout {
+    padding: 0.25rem 0.5rem;
+    border-radius: 5px;
+  }
 
-    .userInfo .username {
-        margin-right: 1rem; 
-    }
-
-    .userInfo .signout {
-        padding: 0.25rem 0.5rem; 
-        border-radius: 5px;
-    }
-
-    .logo {
-        height: 100px;
-    }
-
+  .logo {
+    height: 100px;
+  }
 }
